dotnet: update Magick.NET and SkiaSharp dependencies in demo projects

+++ wrappers/dotnet/ZXingCpp.DemoReader/Program.cs
@@ -15,7 +15,7 @@ public static class MagickImageBarcodeReader
 		if (img.DetermineBitDepth() < 8)
 			img.SetBitDepth(8);
 		var bytes = img.ToByteArray(MagickFormat.Gray);
-		var iv = new ImageView(bytes, img.Width, img.Height, ImageFormat.Lum);
+		var iv = new ImageView(bytes, (int)img.Width, (int)img.Height, ImageFormat.Lum);
 		return BarcodeReader.Read(iv, opts);
 	}
